Dynamic queries 0.1
Our queries have been mostly static. We've
been proxying queries to every other component
we might use, to make sure we've got their data.

Instead, we're now using routing and subqueries
to only query for the data we need.

This is still work in progress. We'll have to fix
all bugs and make it more efficient in at a later time.